Week08 Ex3

+++ week08/ex2.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
 gcc ex2.c -o ex2.out -lm -Wextra -Wpedantic
+if [ $? -ne 0 ]; then
+    echo "Compilation failed!"
+    exit 1
+fi
 ./ex2.out
 rm ex2.out
